+++ web/e2e/project/content/view.spec.ts
@@ -28,7 +28,6 @@ async function itemAdd(page: Page, data: string) {
   await page.getByLabel("Back").click();
 }
 
-// fix this test
 test("View CRUD has succeeded", async ({ page }) => {
   test.slow();
   await page.locator("li").filter({ hasText: "Text" }).locator("div").first().click();
