+++ web/e2e/settings/account/lang.spec.ts
@@ -3,7 +3,6 @@
 import { closeNotification } from "@reearth-cms/e2e/common/notification";
 import { expect, test } from "@reearth-cms/e2e/utils";
 
-// fix this test
 test("Language updating has succeeded", async ({ reearth, page }) => {
   await reearth.goto("/", { waitUntil: "domcontentloaded" });
   await page.getByText(/Account|アカウント/).click();
