+++ web/src/components/molecules/ProjectOverview/index.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { ChangeEvent, useCallback, useEffect, useMemo, useState } from "react";
+import { useMemo } from "react";
 
 import Button from "@reearth-cms/components/atoms/Button";
 import Icon from "@reearth-cms/components/atoms/Icon";
@@ -12,6 +12,7 @@ import { useT } from "@reearth-cms/i18n";
 
 import { Project, SortBy, UpdateProjectInput } from "../Workspace/types";
 
+import useHooks from "./hooks";
 import LicenseTab from "./LicenseTab";
 import ModelsTab from "./ModelsTab";
 import ReadmeTab from "./ReadmeTab";
@@ -33,7 +34,7 @@ type Props = {
   onModelUpdateModalOpen: (model: Model) => Promise<void>;
 };
 
-type ActiveKey = "models" | "readme" | "license";
+export type ActiveKey = "models" | "readme" | "license";
 
 const ProjectOverview: React.FC<Props> = ({
   project,
@@ -52,67 +53,26 @@ const ProjectOverview: React.FC<Props> = ({
   onModelUpdateModalOpen,
 }) => {
   const t = useT();
-  const [activeKey, setActiveKey] = useState<ActiveKey>("models");
 
-  const [activeReadmeTab, setActiveReadmeTab] = useState("edit");
-  const [readmeEditMode, setReadmeEditMode] = useState(false);
-  const [readmeValue, setReadmeValue] = useState("");
-
-  const [activeLicenseTab, setActiveLicenseTab] = useState("edit");
-  const [licenseEditMode, setLicenseEditMode] = useState(false);
-  const [licenseValue, setLicenseValue] = useState("");
-
-  useEffect(() => {
-    if (project?.readme) {
-      setReadmeValue(project.readme);
-    }
-  }, [project?.readme]);
-
-  useEffect(() => {
-    if (project?.license) {
-      setLicenseValue(project.license);
-    }
-  }, [project?.license]);
-
-  const handleReadmeSave = useCallback(async () => {
-    if (!project?.id) return;
-    setActiveReadmeTab("edit");
-    setReadmeEditMode(false);
-    await onProjectUpdate({
-      projectId: project.id,
-      readme: readmeValue,
-    });
-  }, [onProjectUpdate, project?.id, readmeValue]);
-
-  const handleReadmeEdit = useCallback(() => {
-    setReadmeEditMode(true);
-  }, []);
-
-  const handleLicenseSave = useCallback(async () => {
-    if (!project?.id) return;
-    setActiveLicenseTab("edit");
-    setLicenseEditMode(false);
-    await onProjectUpdate({
-      projectId: project.id,
-      license: licenseValue,
-    });
-  }, [onProjectUpdate, project?.id, licenseValue]);
-
-  const handleLicenseEdit = useCallback(() => {
-    setLicenseEditMode(true);
-  }, []);
-
-  const handleReadmeMarkdownChange = useCallback((e: ChangeEvent<HTMLTextAreaElement>) => {
-    setReadmeValue(e.target.value);
-  }, []);
-
-  const handleLicenseMarkdownChange = useCallback((e: ChangeEvent<HTMLTextAreaElement>) => {
-    setLicenseValue(e.target.value);
-  }, []);
-
-  const handleChooseLicenseTemplate = useCallback((value: string) => {
-    setLicenseValue(value);
-  }, []);
+  const {
+    activeKey,
+    setActiveKey,
+    activeReadmeTab,
+    setActiveReadmeTab,
+    readmeEditMode,
+    readmeValue,
+    activeLicenseTab,
+    setActiveLicenseTab,
+    licenseEditMode,
+    licenseValue,
+    handleReadmeSave,
+    handleReadmeEdit,
+    handleLicenseSave,
+    handleLicenseEdit,
+    handleReadmeMarkdownChange,
+    handleLicenseMarkdownChange,
+    handleChooseLicenseTemplate,
+  } = useHooks({ project, onProjectUpdate });
 
   const tabBarExtraContent = useMemo(
     () => (
